+++ src/java/se/idega/idegaweb/commune/school/business/SchoolClassWriter.java
@@ -151,25 +151,25 @@ public class SchoolClassWriter implements MediaWritable {
 	    style.setFont(font);
 
 			int cellRow = 0;
-			HSSFRow row = sheet.createRow((short)cellRow++);
+			HSSFRow row = sheet.createRow(cellRow++);
 			HSSFCell cell = row.createCell((short)0);
 			cell.setCellValue(schoolName);
 			cell.setCellStyle(style);
 			cell = row.createCell((short)1);
 			
-			row = sheet.createRow((short)cellRow++);
+			row = sheet.createRow(cellRow++);
 			cell = row.createCell((short)0);
 			cell.setCellValue(seasonName);
 			cell.setCellStyle(style);
 			
-			row = sheet.createRow((short)cellRow++);
+			row = sheet.createRow(cellRow++);
 			cell = row.createCell((short)0);
 			cell.setCellValue(yearName + " - " + groupName);
 			cell.setCellStyle(style);
 			
-			row = sheet.createRow((short)cellRow++);
+			row = sheet.createRow(cellRow++);
 			
-	    row = sheet.createRow((short)cellRow++);
+	    row = sheet.createRow(cellRow++);
 	    cell = row.createCell((short)0);
 	    cell.setCellValue(iwrb.getLocalizedString("school.name","Name"));
 	    cell.setCellStyle(style);
@@ -194,7 +194,7 @@ public class SchoolClassWriter implements MediaWritable {
 			
 			Iterator iter = students.iterator();
 			while (iter.hasNext()) {
-				row = sheet.createRow((short)cellRow++);
+				row = sheet.createRow(cellRow++);
 				studentMember = (SchoolClassMember) iter.next();
 				student = (User) studentMap.get(new Integer(studentMember.getClassMemberId()));
 				address = userBusiness.getUsersMainAddress(student);
